Switch database from Postgres/Neon to Turso (libSQL/SQLite)
Cheaper, generous free tier, and one driver that runs on both Workers and
Node. Sessions/cache now live in the DB (kv table), so no separate KV service.

- db: schema rewritten in drizzle sqlite-core; createDb via @libsql/client/web
  + drizzle-orm/libsql; drizzle-kit turso dialect; regenerated migration
- api: unified sqlKV(db) KV adapter for both runtimes (dropped Workers KV);
  fixed count(*) and unique-violation detection for SQLite; DATABASE_AUTH_TOKEN
- wrangler: removed kv_namespaces; docs/.env updated for Turso
- verified: in-memory libSQL round-trip (uuid+timestamps, JSON column, kv
  upsert, unique slug constraint) + Node server boots on the libSQL driver

40 tests passing; lint/typecheck/build green.

+++ apps/api/src/routes/auth.ts
@@ -41,7 +41,7 @@ async function parse<T>(c: Context<AppBindings>, schema: z.ZodType<T>): Promise<
 // First-run registration: only allowed when there are zero users. Creates an admin.
 router.post('/register', async (c) => {
   const db = c.get('db');
-  const [row] = await db.select({ count: sql<number>`count(*)::int` }).from(users);
+  const [row] = await db.select({ count: sql<number>`count(*)` }).from(users);
   if ((row?.count ?? 0) > 0) {
     throw errors.forbidden('Registration is closed. Ask an admin to invite you.');
   }
